Isolate routes for faster tests
In a smaller existing application, it might not be worth the effort if it's already structured like this. Locally, when I was benchmarking this, I saved roughly about a half to a full second. Unnecessary micro optimizations can waste valuable development time. Now, what if this was a much larger or more complex application? Absolutely worth it. It can save minutes or even hours depending on the scale of the application. At the end of the day, I feel it's a best practice to structure applications so they can be tested in isolation especially if you're building something new.